% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amp_correlation.R
\name{amp_correlation}
\alias{amp_correlation}
\title{Generates correlation networks from amplicon data.}
\usage{
amp_correlation(data)
}
\arguments{
\item{data}{(required) A phyloseq object including sample data.}

\item{trans}{Transform the raw counts by "sqrt".}

\item{treshold.cor}{Absolute correlation treshold (default: 0.8).}

\item{treshold.pval}{P-value treshold for correlations (default: 0.01).}

\item{ignore.zero}{Remove observations where 1 obeservation has 0 counts from the correlation test (default: F)}

\item{ignore.global.zero}{Remove observations where 1 obeservation has 0 counts from the initial correlation test (default: T)}

\item{scale.seq}{The number of sequences in the pre-filtered samples (default: 20000)}

\item{treshold.abundance}{Average abundance treshold (default: 0)}

\item{treshold.obs}{Minimum number of observations in order to calculate correlations (default: 10).}

\item{label}{Plot taxonomic classifications (e.g. "Genus") instead of points.}

\item{scale.abundance}{Scale the size of nodes by abundance (default: F).}

\item{output}{Either plot or complete (default: "plot").}

\item{treshold.count}{Minimum count in order to use a pair of observations for test of correlation (default: 10)}

\item{scale.size}{Scale the size of the plotted objects (default: 1)}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{tax.empty}{Either "remove" OTUs without taxonomic information, add "best" classification or add the "OTU" name (default: best).}

\item{tax.aggregate}{The taxonomic level that the data should be aggregated to (default: OTU).}

\item{alpha}{Transparency of the plotted points (default: 1).}
}
\value{
A ggplot2 object
}
\description{
A nice long description
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

