% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_export.R
\name{amp_export}
\alias{amp_export}
\title{Export sequences from a phyloseq object}
\usage{
amp_export(data)
}
\arguments{
\item{data}{(required) A phyloseq object.}

\item{file}{Name of the file containing the exported sequences.}

\item{tax}{Add taxonomic strings to the output (default: T).}
}
\description{
Export sequences from a phyloseq object
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

