% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_export_table.R
\name{amp_export_table}
\alias{amp_export_table}
\title{Export otutable from a phyloseq object}
\usage{
amp_export_table(data)
}
\arguments{
\item{data}{(required) A phyloseq object.}

\item{file}{Name of the file containing the exported otutable.}

\item{id}{Name the samples using any variable in the metadata.}

\item{sort.samples}{Vector to sort the samples by.}
}
\description{
Export otutable from a phyloseq object.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

