% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/amp_function.R
\name{amp_function}
\alias{amp_function}
\title{Generate a function table}
\usage{
amp_function(heatmap)
}
\arguments{
\item{heatmap}{(required) A heatmap generated by amp_heatmap.}

\item{functions}{A data frame that associates functions with a genus names (default: MiDAS).}

\item{adjust}{Adjust the whitespace added in the bottom of the plot in mm (default: 10).}

\item{genus.pos}{Specify which part of the name that contains the genus name (default: 1).}
}
\value{
A ggplot2 object.
}
\description{
Generate a function table that is sorted based on an input heatmap.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

