% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_load.R
\name{amp_load}
\alias{amp_load}
\title{Load data as OTU table and convert to a phyloseq object.}
\usage{
amp_load(otutable, metadata)
}
\arguments{
\item{otutable}{(required) A OTU table generated from workflow scripts v.4+.}

\item{metadata}{(required) A metadata file with sample names in first column.}

\item{refseq}{Reference sequences for all OTUs.}

\item{rarefy}{Rarefy all samples to the same sequencing depth.}
}
\value{
A phyloseq object.
}
\description{
Load data as OTU table and convert to a phyloseq object.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

