% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amp_rarecurve.R
\name{amp_rarecurve}
\alias{amp_rarecurve}
\title{Calculate rarefaction curve for each sample.}
\usage{
amp_rarecurve(data)
}
\arguments{
\item{data}{(required) A phyloseq object.}

\item{step}{Step size for sample sizes in rarefaction curves (default: 100).}

\item{ylim}{vector of y-axis limits.}

\item{xlim}{vector of x-axis limits.}

\item{label}{Label rarefaction curves (default: F).}

\item{color}{Color lines by metadata.}
}
\description{
Calculate rarefaction curve for each sample using the vegan rarecurve function directly from a phyloseq object.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

