% Generated by roxygen2 (4.0.1): do not edit by hand
\name{amp_rename}
\alias{amp_rename}
\title{Used for cleaning and renaming taxonomy in phyloseq objects}
\usage{
amp_rename(data)
}
\arguments{
\item{data}{(required) A phyloseq object.}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{tax.empty}{Either "remove" OTUs without taxonomic information at X level, with "best" classification or add the "OTU" name (default: best).}

\item{tax.level}{The taxonomic level to remove OTUs with empty taxonomy, only used when tax.empty = "remove" (default: Genus).}
}
\value{
A phyloseq object with cleaned and renamed taxonomy.
}
\description{
Used internally in other ampvis functions.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

