% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/amp_stability.R
\name{amp_stability}
\alias{amp_stability}
\title{Calculate similarity over time}
\usage{
amp_stability(data)
}
\arguments{
\item{data}{(required) A phyloseq object including sample data.}

\item{date}{(required) The name of the variable that stores the date information.}

\item{group}{Split the dataset into selected groups using a metadata variable.}

\item{plot.type}{Either "time" or "delta" (default: "time").}

\item{plot.theme}{Chose different standard layouts choose from "normal" or "clean" (default: "clean").}

\item{output}{Either "plot" or "complete" (default: "plot").}
}
\description{
A nice long description.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

