% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_stats.R
\name{amp_stats}
\alias{amp_stats}
\title{Calculate basic statistics for each sample}
\usage{
amp_stats(data)
}
\arguments{
\item{data}{(required) A phyloseq object.}

\item{measure}{Alpha-diversity measures to be included (default:observed).}
}
\description{
Calculate basic statistics for each sample and combine it with metadata.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

