% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_test_species.R
\name{amp_test_species}
\alias{amp_test_species}
\title{Tests if there is a significant difference in abundance between selected conditions}
\usage{
amp_test_species(data, design)
}
\arguments{
\item{data}{(required) A phyloseq object including sample data.}

\item{group}{(required) The group to test against.}

\item{tax.aggregate}{Group data at specific taxonomic level (default: "OTU").}

\item{sig}{Significance treshold (default: 0.01).}

\item{fold}{Log2fold filter default for displaying significant results (default: 0)}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{tax.empty}{Either "remove" OTUs without taxonomic information at X level, with "best" classification or add the "OTU" name (default: best).}

\item{label}{Label the significant entries with tax.display (default:F).}

\item{plot.type}{Either "boxplot" or "point" (default: point)}

\item{plot.show}{Display the X most significant results.}

\item{plot.point.size}{The size of the plotted points.}

\item{plot.theme}{Chose different standard layouts choose from "normal" or "clean" (default: "normal").}

\item{parallel}{To run DESeq2 in parallel mode (default: F)}

\item{adjust.zero}{Keep 0 abundances in ggplot2 median calculations by adding a small constant to these.}

\item{tax.display}{Display additional taxonomic levels in the plot output e.g. "Genus".}
}
\value{
A p-value for each comparison.
}
\description{
A nice long description
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

