\docType{package}
\name{ampvis}
\alias{ampvis}
\alias{ampvis-package}
\title{ampvis: amplicon visualisation}
\description{
Tools for visualisation of amplicon data
}
\section{Individual functions}{
  \itemize{ \item \code{\link{amp_heatmap}}: Generate a
  heatmap from amplicon data.  \item
  \code{\link{amp_rabund}}: Generate a rank abundance curve
  from amplicon data.  \item \code{\link{amp_ordinate}}:
  Generate ordination plots from amplicon data. }
}

