\docType{data}
\name{data}
\alias{data}
\title{Tests if there is a significant difference in abundance between selected conditions}
\format{\preformatted{Formal class 'phyloseq' [package "phyloseq"] with 5 slots
  ..@ otu_table:Formal class 'otu_table' [package "phyloseq"] with 2 slots
  .. .. ..@ .Data        : num [1:380, 1:28] 0 5 0 0 1 0 0 0 0 0 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:380] "denovo38" "denovo256" "denovo394" "denovo433" ...
  .. .. .. .. ..$ : chr [1:28] "16S-AMP-1147" "16S-AMP-1149" "16S-AMP-1156" "16S-AMP-1113" ...
  .. .. ..@ taxa_are_rows: logi TRUE
  ..@ tax_table:Formal class 'taxonomyTable' [package "phyloseq"] with 1 slots
  .. .. ..@ .Data: chr [1:380, 1:6] "k__Bacteria" "k__Bacteria" "k__Bacteria" "k__Bacteria" ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:380] "denovo38" "denovo256" "denovo394" "denovo433" ...
  .. .. .. .. ..$ : chr [1:6] "Kingdom" "Phylum" "Class" "Order" ...
  ..@ sam_data :'data.frame':	28 obs. of  14 variables:
Formal class 'sample_data' [package "phyloseq"] with 4 slots
  .. .. ..@ .Data    :List of 14
  .. .. .. ..$ : Factor w/ 28 levels "16S-AMP-1106",..: 11 13 14 6 21 9 8 12 16 15 ...
  .. .. .. ..$ : Factor w/ 7 levels "Bjergmarken",..: 6 6 2 5 4 1 1 6 2 2 ...
  .. .. .. ..$ : Factor w/ 4 levels "2013-02-04","2013-05-06",..: 2 4 2 1 3 3 2 3 4 3 ...
  .. .. .. ..$ : int [1:28] 2013 2013 2013 2013 2013 2013 2013 2013 2013 2013 ...
  .. .. .. ..$ : Factor w/ 4 levels "Fall","Spring",..: 2 1 2 4 3 3 2 3 1 3 ...
  .. .. .. ..$ : Factor w/ 2 levels "N","Y": 2 2 1 1 1 2 2 2 1 1 ...
  .. .. .. ..$ : Factor w/ 2 levels "N","Y": 1 1 1 2 2 1 1 1 1 1 ...
  .. .. .. ..$ : Factor w/ 1 level "Y": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ : Factor w/ 2 levels "N","Y": 2 2 2 2 1 2 2 2 2 2 ...
  .. .. .. ..$ : Factor w/ 2 levels "N","Y": 1 1 1 1 2 1 1 1 1 1 ...
  .. .. .. ..$ : Factor w/ 2 levels "N","Y": 2 2 1 1 2 2 2 2 1 1 ...
  .. .. .. ..$ : Factor w/ 5 levels "0.1","0.2","0.3",..: 1 1 4 2 3 2 2 1 4 4 ...
  .. .. .. ..$ : Factor w/ 2 levels "N","Y": 2 2 1 1 2 1 1 2 1 1 ...
  .. .. .. ..$ : Factor w/ 3 levels "Acetic acid and protein",..: 2 2 3 3 1 3 3 2 3 3 ...
  .. .. ..@ names    : chr [1:14] "SampleID" "Plant" "Date" "Year" ...
  .. .. ..@ row.names: chr [1:28] "16S-AMP-1147" "16S-AMP-1149" "16S-AMP-1156" "16S-AMP-1113" ...
  .. .. ..@ .S3Class : chr "data.frame"
  ..@ phy_tree : NULL
  ..@ refseq   : NULL
}}
\usage{
amp_test(data, group)
}
\arguments{
  \item{data}{(required) A phyloseq object including sample
  data.}

  \item{group}{(required) A variable from the associated
  sample data to group samples by.}

  \item{tax.aggregate}{Group data at specific taxonomic
  level (defaul: "Genus").}

  \item{tax.clean}{Replace the phylum Proteobacteria with
  the respective Classes instead (default: T).}

  \item{tax.empty}{Either "remove" OTUs without taxonomic
  information or "rename" with OTU ID (default: rename).}

  \item{scale.seq}{The number of sequences in the
  pre-filtered samples (default: 20000)}
}
\value{
A p-value for each comparison.
}
\description{
A nice long description
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
\keyword{datasets}

