% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/amp_convert.R
\name{amp_convert}
\alias{amp_convert}
\title{Convert a phyloseq object to a list of dataframes.}
\usage{
amp_convert(data)
}
\arguments{
\item{data}{(required) A phyloseq object.}
}
\value{
A list of dataframes.
}
\description{
Convert a phyloseq object to a list of dataframes.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

