% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amp_core.R
\name{amp_core}
\alias{amp_core}
\title{Generates a ggplot2 style core community plots}
\usage{
amp_core(data)
}
\arguments{
\item{data}{(required) A phyloseq object (or a list).}

\item{group}{Group the data based on a sample variable (default: "Sample").}

\item{scale.seq}{The number of sequences in the pre-filtered samples (default: 10000)}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{tax.empty}{Either "remove" OTUs without taxonomic information, add "best" classification or add the "OTU" name (default: best).}

\item{plot.type}{Either core or frequency (default: frequency).}

\item{output}{Either plot or complete (default: "plot").}

\item{tax.aggregate}{The taxonomic level that the data should be aggregated to (default: OTU).}

\item{weight}{Weight the frequency by abundance (default: T).}

\item{abund.treshold}{Treshold for considering something abundant in percent (default: 0.1).}
}
\value{
A ggplot2 object
}
\description{
A nice long description
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

