% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amp_ordinate.R
\name{amp_ordinate}
\alias{amp_ordinate}
\title{Generates a ggplot2 style ordinate plot of amplicon data.}
\usage{
amp_ordinate(data)
}
\arguments{
\item{data}{(required) A phyloseq object including sample data.}

\item{scale}{A variable from the associated sample data to scale the abundance by.}

\item{trans}{Transform the raw counts, currently supports "none" or "sqrt" (default: "sqrt").}

\item{ordinate.type}{Either PCA or NMDS (default: "PCA").}

\item{ncomp}{The number of principal components to extract if using PCA (default: 5)}

\item{plot.x}{Variable to plot on the x-axis if using PCA (default: "PC1")}

\item{plot.y}{Variable to plot on the y-axis if using PCA (default: "PC2")}

\item{plot.color}{Color the points by a sample variable.}

\item{plot.color.order}{Order the groups used for coloring by a vector.}

\item{plot.point.size}{Size of the plotted sample points (default: 3)}

\item{plot.shape}{Shape points by a sample variable.}

\item{plot.species}{Plot loadings as points (default: F)}

\item{plot.nspecies}{Plot the n most extreme species with their taxonomic classification (default: 0).}

\item{plot.nspecies.tax}{Taxonomic level used in plot.nspecies (default: "Genus").}

\item{plot.label}{Label points using a sample variable.}

\item{plot.group}{Uses plot.color and groups samples by either "chull" or "centroid".}

\item{plot.group.label}{Add label based on the centroid of the specified group.}

\item{envfit.factor}{A vector of factor variables from the sample data used for envfit to the model.}

\item{envfit.numeric}{A vector of numerical variables from the sample data used for envfit to the model.}

\item{envfit.significant}{The significance treshold for displaying envfit parameters (default: 0.01).}

\item{envfit.resize}{Scale the size of the numeric arrows (default: 1).}

\item{envfit.color}{Color of the envfit text on the plot (default: "darkred").}

\item{envfit.textsize}{Size of the envfit text on the plot (default: 3).}

\item{envfit.show}{Show the results on the plot (default: T).}

\item{tax.empty}{Option to add "best" classification or just the "OTU" name to each "OTU" (default: best).}

\item{output}{Either plot or complete (default: "plot").}

\item{constrain}{Constrain the PCA by a sample variable.}

\item{scale.species}{Rescale the plotted loadings to maximise visability (default: F).}

\item{trajectory}{Connects points based on a sample variable e.g. date.}

\item{trajectory.group}{Split the trajectory by a group.}

\item{plot.group.label.size}{Text size of the labels.}

\item{plot.theme}{Chose different standard layouts choose from "normal" or "clean" (default: "normal").}
}
\value{
A ggplot2 object
}
\description{
A nice long description
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

