% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amp_rabund.R
\name{amp_rabund}
\alias{amp_rabund}
\title{Generates a ggplot2 style rank abundance plot of the most abundant OTUs}
\usage{
amp_rabund(data)
}
\arguments{
\item{data}{(required) A phyloseq object.}

\item{group}{Group the data based on a sample variable.}

\item{order.group}{A vector defining the order of groups.}

\item{tax.show}{The number of taxa to show or a vector of taxa names (default: 50).}

\item{scale.seq}{The number of sequences in the pre-filtered samples (default: 10000)}

\item{tax.clean}{Replace the phylum Proteobacteria with the respective Classes instead (default: T).}

\item{plot.type}{Either point, boxplot or curve (default: boxplot).}

\item{plot.log}{Log10 scale the data (default: F)}

\item{output}{Either plot or complete (default: "plot").}

\item{tax.add}{Additional taxonomic levels to display for each entry (default: Phylum)}

\item{tax.aggregate}{The taxonomic level that the data should be aggregated to (defualt: OTU)}

\item{tax.empty}{Either "remove" OTUs without taxonomic information, "rename" with best classification or add the "OTU" name (default: rename).}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{point.size}{Size of points (default: 2).}

\item{plot.flip}{Flip the axis of the plot (default: F).}

\item{sort.by}{Sort the boxplot by either Median, Mean or Total (default = "Median")}

\item{adjust.zero}{Keep 0 abundances in ggplot2 median calculations by adding a small constant to these.}

\item{plot.theme}{Chose different standard layouts choose from "normal" or "clean" (default: "normal").}
}
\value{
A ggplot2 object
}
\description{
A nice long description
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

