% Generated by roxygen2 (4.0.1): do not edit by hand
\name{amp_test_cluster}
\alias{amp_test_cluster}
\title{Tests if there is a significant difference between samples.}
\usage{
amp_test_cluster(data, group)
}
\arguments{
\item{data}{(required) A phyloseq object including sample data.}

\item{group}{(required) The group variables to be tested.}

\item{method}{The distance measure (default: "bray").}

\item{plot.label}{A vector specifying the sample data to be used as labels.}

\item{plot.color}{A vector specifying the sample data to be used as colors.}
}
\value{
A number of things.
}
\description{
A nice long description
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

