% Generated by roxygen2 (4.0.1): do not edit by hand
\name{amp_test_species}
\alias{amp_test_species}
\title{Tests if there is a significant difference in abundance between selected conditions}
\usage{
amp_test_species(data, design)
}
\arguments{
\item{data}{(required) A phyloseq object including sample data.}

\item{group}{(required) The group to test against.}

\item{sig}{Significance treshold (default: 0.01).}

\item{fold}{Log2fold filter default for displaying significant results (default: 0)}

\item{tax.aggregate}{Group data at specific taxonomic level (default: "OTU").}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{tax.empty}{Either "remove" OTUs without taxonomic information at X level, with "best" classification or add the "OTU" name (default: best).}

\item{tax.display}{Display additional taxonomic levels in the plot output e.g. "Genus".}

\item{label}{Label the significant entries with tax.display (default:F).}

\item{plot.type}{Either "boxplot" or "point" (default: point)}

\item{plot.show}{Display the X most significant results.}

\item{plot.point.size}{The size of the plotted points.}
}
\value{
A p-value for each comparison.
}
\description{
A nice long description
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}

